a shared component handling the form with react-hook-form i now shared between TaskConfigureModal and TaskCreateModal, whereas TaskCreateModal works, but TaskConfigureModal still needs to be configured to be used in the constellation
