+++ todo_mern_client/src/components/Modal/components/NewCategoryForm/newcategoryform.module.scss
@@ -1,0 +1,76 @@
+.form {
+  display: none;
+  grid-template-columns: repeat(20, 1fr);
+  grid-template-rows: 1fr 1fr;
+  position: absolute;
+  grid-row: 10/13;
+  justify-self: center;
+  width: 80%;
+  z-index: 105;
+  height: 7rem;
+  padding: .5rem 1rem;
+  margin: 0 1rem;
+  align-items: center;
+  /* gap: 1rem; */
+  background: var(--dark-blue);
+  /* border: 1px solid black; */
+  border-radius: 0.5rem;
+  /* box-shadow: 0 0 10px 0 var(--pink); */
+  box-shadow: 0 0 10px 0 white;
+
+
+  &Visible {
+    display: grid;
+  }
+
+  &__closeButton {
+    position: absolute;
+    top: -10px;
+    left: -10px;
+    width: 1.5rem;
+    height: 1.5rem;
+    border-radius: 50%;
+    font-size: 1.5rem;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+  @-moz-document url-prefix() {
+    &__closeButton {
+      padding-bottom: 4.5px;
+    }
+  }
+
+  &__inputName {
+    width: 100%;
+    grid-column: 1/16;
+    height: 2rem;
+    background: var(--light-blue);
+    border: none;
+    color: var(--text-color);
+    text-align: center;
+    border-radius: 0.5rem;
+  }
+
+  &__colorPicker {
+    grid-column: 17/21;
+    justify-self: center;
+    height: 1.8rem;
+    width: 1.8rem;
+    border-radius: 50%;
+    border: none;
+    right: 0;
+  }
+
+  &__submitButton {
+    background: var(--pink);
+    border: 1px solid var(--light-blue);
+    border-radius: 0.5rem;
+    color: var(--text-color);
+    z-index: 100;
+    grid-column: 1/21;
+    grid-row: 2/3;
+    height: 2rem;
+  }
+}
