+++ todo_mern_client/src/components/Modal/taskmodal.module.scss
@@ -1,0 +1,195 @@
+  /*TaskModal variables*/
+  $taskModalBorderRamius: 1rem;
+
+  /*exitButton variables*/
+  $exitButtonSize: 1.9rem;
+  $top-left-distance: 0.8rem;
+
+  @keyframes showModalAnim {
+    20% {
+      right: 20%;
+    }
+
+    60% {
+      right: -1%;
+    }
+
+    100% {
+      /*5% to land in the middle(corresponding to its width)*/
+      right: 5%;
+    }
+  }
+
+  @keyframes hideModalAnim {
+    0% {
+      right: 0;
+    }
+
+    20% {
+      right: 30%;
+    }
+
+    99% {
+      right: -100%;
+    }
+
+    100% {
+      visibility: hidden;
+    }
+
+  }
+
+  .modalHide {
+    animation: hideModalAnim 0.5s forwards;
+  }
+
+  .modalVisible {
+    animation: showModalAnim 0.5s forwards;
+  }
+
+  .TaskCreateModal {
+    position: absolute;
+    background: var(--dark-blue);
+    height: 90svh;
+    width: 90vw;
+    z-index: 1000;
+    align-self: center;
+    border-radius: $taskModalBorderRamius;
+    /* box-shadow: 1px 2px 14px 5px black; */
+    right: -100%;
+    border: 2px solid var(--light-blue);
+    display: grid;
+    grid-template-rows: repeat(20, 1fr);
+
+    &Hide {
+      animation: hideModalAnim 0.5s forwards;
+    }
+
+    &Visible {
+      animation: showModalAnim 0.5s forwards;
+    }
+
+    &__exitButton {
+      height: $exitButtonSize;
+      width: $exitButtonSize;
+      border-radius: 100%;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      font-size: 2rem;
+      border: none;
+      position: absolute;
+      top: $top-left-distance;
+      left: $top-left-distance;
+    }
+
+    /*The exit(&times;) sign wouldnt center properly in firefox*/
+    @-moz-document url-prefix() {
+      &__exitButton {
+        padding-bottom: 4.5px;
+      }
+    }
+  }
+
+
+
+  .form {
+    display: grid;
+    grid-template-rows: 1fr 3fr 1fr 1fr 1fr;
+    gap: 0.5rem;
+    padding: 1rem;
+    grid-row: 2/21;
+    position: relative;
+
+    &__inputWrapper {
+      display: flex;
+      flex-direction: column;
+      gap: 0.5rem;
+    }
+
+    &__label {
+      color: var(--text-color);
+      display: flex;
+      justify-content: center;
+      font-size: 1.2rem;
+    }
+
+    &__input {
+      background: var(--light-blue);
+      border: none;
+      border-radius: 0.5rem;
+      padding: 0.5rem;
+      color: var(--text-color);
+      height: 100%;
+      color: var(--text-color);
+      font-size: 1rem;
+      width: 100%;
+    }
+
+    &__select {
+      background: var(--light-blue);
+      border: none;
+      border-radius: 0.5rem;
+      padding: 0.5rem;
+      color: var(--text-color);
+    }
+
+    &__createNewCategoryButton {
+      grid-column: 9/11;
+      background: var(--pink);
+      border: 1px solid var(--light-blue);
+      border-radius: 0.5rem;
+      color: var(--text-color);
+      z-index: 100;
+    }
+
+    &__submit {
+      background: var(--pink);
+      color: white;
+      border: none;
+      border-radius: 0.5rem;
+      padding: 0.5rem;
+      font-size: 1.2rem;
+      height: 3.4rem;
+      align-self: center;
+      cursor: pointer;
+      margin-top: 1rem;
+    }
+  }
+
+
+
+  .inputWrapperCategory {
+    display: grid;
+    grid-template-columns: repeat(10, 1fr);
+    position: relative;
+
+    &__label {
+      grid-column: 1/11;
+    }
+
+    &__select {
+      grid-column: 1/9;
+    }
+
+    &__button {
+      grid-column: 9/11;
+      background: var(--pink);
+      border: 1px solid var(--light-blue);
+      border-radius: 0.5rem;
+      color: var(--text-color);
+      z-index: 98;
+    }
+  }
+
+
+
+  .hideBackgroundWhenNewCategoryIsOpen {
+    position: absolute;
+    height: 100%;
+    width: 100%;
+    border-radius: $taskModalBorderRamius;
+    background: black;
+    opacity: 0.8;
+    z-index: 98;
+  }
